Move repositories and queries into their own packages
