+++ src/test/java/com/example/whereabouts/humanresources/EmployeeTestDataService.java
@@ -5,8 +5,8 @@ import com.example.whereabouts.common.EmailAddress;
 import com.example.whereabouts.common.Gender;
 import com.example.whereabouts.common.PhoneNumber;
 import com.example.whereabouts.common.address.*;
-import com.example.whereabouts.humanresources.internal.EmployeeRepository;
-import com.example.whereabouts.humanresources.internal.EmploymentDetailsRepository;
+import com.example.whereabouts.humanresources.repository.EmployeeRepository;
+import com.example.whereabouts.humanresources.repository.EmploymentDetailsRepository;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
 import org.springframework.stereotype.Component;
