+++ src/test/java/com/example/whereabouts/projects/repository/ProjectRepositoryTest.java
@@ -1,0 +1,30 @@
+package com.example.whereabouts.projects.repository;
+
+import com.example.whereabouts.IntegrationTest;
+import com.example.whereabouts.projects.ProjectData;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@IntegrationTest
+class ProjectRepositoryTest {
+
+    @Autowired
+    ProjectRepository repository;
+
+    static ProjectData createProjectData() {
+        return new ProjectData("Name", "Description");
+    }
+
+    @Test
+    void insert_and_get_include_all_properties() {
+        var originalData = createProjectData();
+        var id = repository.insert(originalData);
+
+        var retrieved = repository.findById(id).orElseThrow();
+        assertThat(retrieved.id()).isEqualTo(id);
+        assertThat(retrieved.version()).isEqualTo(1);
+        assertThat(retrieved.data()).isEqualTo(originalData);
+    }
+}
