+++ src/test/java/com/example/whereabouts/security/UserTestDataService.java
@@ -1,6 +1,6 @@
 package com.example.whereabouts.security;
 
-import com.example.whereabouts.security.internal.UserRepository;
+import com.example.whereabouts.security.repository.UserRepository;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
